move isValidParameter and getArgumentType to @wdio/utils
