+++ packages/wdio-utils/src/index.js
@@ -1,12 +1,14 @@
 import initialisePlugin from './initialisePlugin'
 import initialiseServices from './initialiseServices'
 import webdriverMonad from './monad'
-import { commandCallStructure, isFunctionAsync } from './utils'
+import { commandCallStructure, isValidParameter, getArgumentType, isFunctionAsync } from './utils'
 
 export {
     initialisePlugin,
     initialiseServices,
     isFunctionAsync,
     webdriverMonad,
-    commandCallStructure
+    commandCallStructure,
+    isValidParameter,
+    getArgumentType
 }
